feat(dataset): 优化数据集搜索功能
- 添加 paragraphSearch 方法，支持按问题搜索数据集段落
- 实现数据集段落的处理和格式化
- 添加数据集验证接口
- 优化数据集相关实体和 DTO 结构

+++ src/main/java/com/tarzan/maxkb4j/module/application/chatpipeline/step/searchdatasetstep/impl/SearchDatasetStep.java
@@ -33,7 +33,7 @@ public class SearchDatasetStep extends ISearchDatasetStep {
         HitTestDTO hitTestDTO=new HitTestDTO();
         hitTestDTO.setQuery_text(execProblemText);
         hitTestDTO.setSearch_mode(datasetSetting.getString("search_mode"));
-        hitTestDTO.setSimilarity(datasetSetting.getDouble("similarity"));
+        hitTestDTO.setSimilarity(datasetSetting.getFloat("similarity"));
         hitTestDTO.setTop_number(datasetSetting.getInteger("top_n"));
         List<ParagraphVO> paragraphList= embeddingService.paragraphSearch(application.getDatasetIdList(),hitTestDTO);
         System.out.println("search 耗时 "+(System.currentTimeMillis()-startTime)+" ms");
