+++ src/main/java/com/tarzan/maxkb4j/module/dataset/dto/HitTestDTO.java
@@ -2,13 +2,11 @@ package com.tarzan.maxkb4j.module.dataset.dto;
 
 import lombok.Data;
 
-import java.util.UUID;
-
 @Data
 public class HitTestDTO {
     private String  query_text;
     private String  search_mode;
-    private Double  similarity;
+    private Float  similarity;
     private Integer  top_number;
     private Boolean  problem_optimization;
     private String reranker_model_id;
